Implemented user reset password dialog
