+++ client/components/Dialog/actions.js
@@ -1,6 +1,7 @@
 export const DIALOG_IDS = {
   LOGIN: 'login',
   REGISTER: 'register',
+  RESET_PASSWORD: 'resetPassword',
 }
 
 export const showDialog = (id) => {
@@ -20,3 +21,4 @@ export const closeDialog = () => ({
 
 export const openLoginDialog = () => showDialog(DIALOG_IDS.LOGIN)
 export const openRegisterDialog = () => showDialog(DIALOG_IDS.REGISTER)
+export const openResetPasswordDialog = () => showDialog(DIALOG_IDS.RESET_PASSWORD)
