--- client/pages/UserResetPassword/ResendVerificationButton.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const ResendVerificationButton = styled.div`
-  margin-top: 20px;
-  text-decoration: underline;
-  cursor: pointer;
-`
-
-export default ResendVerificationButton
